requirements.txt
Flask==2.2.5
librosa==0.10.1
matplotlib==3.5.3
numpy==1.21.6
python-dotenv==0.21.1
Requests==2.31.0
scikit_learn==1.0.2
torch==1.12.1+cu113
transformers==4.30.2
